+++ playground/README.md
@@ -1,0 +1,15 @@
+### Playground
+
+In order to run webpack in 'watch' mode:
+
+`npm run start`
+
+Run node package from a JS file:
+
+`npm run playground:js`
+
+Run node package from a TS file (assuming ts-node installed globally):
+
+`npm run playground:ts`
+
+Run browser package from a HTML page: Open client.html with your browser.
